requirements.txt
pandas
requests
Pillow
cohere